Notification sending added
